re-enable gdb debugging on the validationslave using core dumps

+++ test/validation/run.sh
@@ -13,6 +13,14 @@ if [ ! -x "$1" ]; then
 	exit 1
 fi
 
+if [ "$(cat /proc/sys/kernel/core_pattern)" != "core" ]; then
+	echo "Please set /proc/sys/kernel/core_pattern to core"
+	exit 1
+fi
+
+# enable core dumps
+ulimit -c unlimited
+
 RUNCLIENT=test/validation/run-client.sh
 
 if [ ! -x $RUNCLIENT ]; then
@@ -20,13 +28,6 @@ if [ ! -x $RUNCLIENT ]; then
 	exit 1
 fi
 
-GDBCMDS=$(mktemp)
-(
-	echo file $1
-	echo run $2 $3 $4 $5 $6 $7 $8 $9
-	echo bt full
-	echo quit
-)>$GDBCMDS
 
 # limit to 1GB RAM
 ulimit -v 1000000
@@ -42,30 +43,50 @@ $RUNCLIENT $1 &
 PID_CLIENT=$!
 
 set +e #temp disable abort on error
-#gdb -batch -return-child-result -x $GDBCMDS &
 $@ &
-
 PID_HOST=$!
 
 # auto kill host after 15mins
 #sleep 900 && kill -9 $PID_HOST &
 
+echo waiting for host to exit
 # store exit code
 wait $PID_HOST
 EXIT=$?
-set -e
 
-# cleanup
-rm -f $GDBCMDS
 
+echo waiting for client to exit
 # get spring client process exit code / wait for exit
 wait $PID_CLIENT
 EXITCHILD=$?
+
+#reenable abbort on error
+set -e
+
+echo Server and client exited
+
+COREFILE=$HOME/.spring/core
+if [ -s $COREFILE ]; then
+	echo Core file found, creating backtrace
+	GDBCMDS=$(mktemp)
+	(
+		echo file $1
+		echo core-file $HOME/.spring/core
+		echo bt full
+		echo quit
+	)>$GDBCMDS
+	gdb -batch -x $GDBCMDS
+	cat $GDBCMDS
+	# cleanup
+	rm -f $GDBCMDS
+fi
+
 # wait for client to exit
 if [ $EXITCHILD -ne 0 ];
 then
-	echo "Error: Spring-client exited with error code $EXITCHILD"
+	echo Client exited with $EXITCHILD
 	exit $EXITCHILD
 fi
+
 exit $EXIT
